+++ CHANGES.md
@@ -1,4 +1,4 @@
-## UNRELEASED
+## v0.10.0 (2014-04-02)
 
 - Issue warning instead of failing when installed version of SQLAlchemy isn't compatible with `alchy.Query`'s loading API (i.e. missing `sqlalchemy.orm.strategy_options.Load`). This allows `alchy` to be used with earlier versions of SQLAlchemy at user's own risk.
 - Add `alchy.search` module which provides compatible search functions for `ModelBase.__advanced_search__` and `ModelBase.__simple_search__`.